% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_addition.R
\name{matrix_addition}
\alias{matrix_addition}
\title{Sum a matrix}
\usage{
matrix_addition(numeric.matrix)
}
\arguments{
\item{numeric.matrix}{Matrix to be summed}
}
\value{
Integer
}
\description{
Sum a matrix
}
